refactor: extract duplicated codes
